修改amount类型，支持小数

+++ demo_back/back/pojo/src/main/java/com/example/dto/UserBillDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,7 +16,7 @@ public class UserBillDTO {
     private Long id;
     private Long userId; //关联用户
     private Long categoryId; //关联分类
-    private Long amount; //金额
+    private BigDecimal amount; //金额
     private int type; //类型 1.收入 2.支出
     private String remark; //备注
 }
