+++ demo_back/back/pojo/src/main/java/com/example/entity/SumStatistics.java
@@ -5,13 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 // 统计结果POJO（用于接收一次查询的多维度总和）
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class SumStatistics {
-    private Long income;    // 收入总和
-    private Long expense;   // 支出总和
-    private Long transfer;  // 转账总和
+    private BigDecimal income;    // 收入总和
+    private BigDecimal expense;   // 支出总和
+    private BigDecimal transfer;  // 转账总和
 }
